requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
psycopg2-binary==2.9.5
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
